Move tensor_utils.h into micro/kernels/ directory
