+++ tensorflow/lite/micro/kernels/tensor_utils.h
@@ -1,4 +1,4 @@
-/* Copyright 2017 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2022 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -12,8 +12,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#ifndef TENSORFLOW_LITE_KERNELS_INTERNAL_TENSOR_UTILS_H_
-#define TENSORFLOW_LITE_KERNELS_INTERNAL_TENSOR_UTILS_H_
+#ifndef TENSORFLOW_LITE_MICRO_KERNELS_TENSOR_UTILS_H_
+#define TENSORFLOW_LITE_MICRO_KERNELS_TENSOR_UTILS_H_
 
 #include <cmath>
 #include <cstdint>
@@ -205,4 +205,4 @@ inline void ApplyActivationToVector(const float* vector, int v_size,
 }  // namespace tensor_utils
 }  // namespace tflite
 
-#endif  // TENSORFLOW_LITE_KERNELS_INTERNAL_TENSOR_UTILS_H_
+#endif  // TENSORFLOW_LITE_MICRO_KERNELS_TENSOR_UTILS_H_
